refactor: update workflows quickstart after refactor
